Extract non-ckpt util function
